+++ server/Gragh/DTO/DtoUser.cs
@@ -10,7 +10,7 @@ namespace DTO
 {
     public class DtoUser
     {
-        public long Code { get; set; }
+        public int Code { get; set; }
         public string firstName { get; set; }
         public string lastName { get; set; }
         public string userName { get; set; }
